% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FACTORIAL_FUN.R
\name{func_test}
\alias{func_test}
\title{Two-sample multiple-direction log rank test}
\usage{
func_test(
  formula,
  event = "event",
  data = NULL,
  nperm = 10000,
  alpha = 0.05,
  cross = TRUE,
  nested.levels.unique = FALSE,
  rg = list(c(0, 0))
)
}
\arguments{
\item{formula}{A model \code{formula} object. The left hand side contains the time variable and the right
hand side contains the factor variables of interest. An interaction term must be specified.}

\item{event}{Name of the response variable}

\item{data}{A data.frame, list or environment containing the variables \code{time},
\code{event} (with values 0 for censored and 1 for uncensored) and \time{group}.}

\item{nperm}{The number of permutations used for calculating the permuted p-value.
The default option is 10000.}

\item{alpha}{A number specifying the significance level; the default is 0.05.}

\item{cross}{logical. Should the weight correspondng to crossing hazards be included?
The default is \code{TRUE}.}

\item{nested.levels.unique}{A logical specifying whether the levels of the nested factor(s) are labeled uniquely or not.
Default is FALSE, i.e., the levels of the nested factor are the same for each level of the main factor.}

\item{rg}{A list (or \code{NULL}) containing the exponents \code{c(r, g)} of the directions
 \eqn{w(x) = x^r (1-x)^g}. Both exponents need to be natural numbers including 0.
Default is \code{list( c(0, 0) )} corresponding to proportional hazards.}
}
\value{
A
}
\description{
The mdir.logrank function calculates the p-values of the multiple-direction logrank test based
on the \eqn{\chi^2}-approximation and the permutation approach.
}
\details{
The package provides the multiple-direction logrank statistic for
  the two sample testing problem withing right-censored survival data. Directions
  of the form w(x) = 1 - 2x (\code{cross = TRUE}) and w(x) = x^r * (1-x)^g for natural numbers
  r,g (including 0) can be specified.
  The multiple-direction logrank test needs linearly independent directions.
  A check for this is implement. If the directions chosen by the user are
  linearly independent then a subset consisting of linearly independent directions
  is selected automatically.

  The \code{mdir.logrank} function returns the test statistic as well as two
  corresponding p-values: the first is based on a \eqn{chi^2} approximation and
  the second one is based on a permutation procedure.
}
\examples{
library(coin)
data(GTSG)
out <- mdir.logrank(data = GTSG)

## Detailed informations:
summary(out)

}
\references{
Ditzhaus, M., Friedrich, S. (2018). Titel und so (Theory)

Ditzhaus, M., Friedrich, S. (2018). Titel und so (practical paper)
}
