% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGUI.R
\name{GFDsurvGUI}
\alias{GFDsurvGUI}
\title{A shiny app for the package GFDsurv}
\usage{
GFDsurvGUI()
}
\description{
This function provides a shiny app for calculating
CASANOVA, medSANOVA and copSANOVA test statistics and respective p-values.
}
