% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casanova.R
\name{casanova}
\alias{casanova}
\title{CASANOVA: Cumulative Aalen survival analyis-of-variance}
\usage{
casanova(
  formula,
  event = "event",
  data = NULL,
  nperm = 1999,
  alpha = 0.05,
  cross = TRUE,
  nested.levels.unique = FALSE,
  rg = list(c(0, 0))
)
}
\arguments{
\item{formula}{A model \code{formula} object. The left hand side contains the time variable and the right
hand side contains the factor variables of interest. An interaction term must be
specified.}

\item{event}{The name of censoring status indicator with values 0=censored and
1=uncensored.
The default choice is "event"}

\item{data}{A data.frame, list or environment containing the variables in formula
and the censoring status
indicator. Default option is \code{NULL}.}

\item{nperm}{The number of permutations used for calculating the permuted p-value.
The default option is 1999.}

\item{alpha}{A number specifying the significance level; the default is 0.05.}

\item{cross}{logical. Should the crossing weight w(x) = 1 - 2x be included?
The default is \code{TRUE}.}

\item{nested.levels.unique}{A logical specifying whether the levels of the nested
factor(s) are labeled uniquely or not.
 Default is FALSE, i.e., the levels of the nested factor are the same for each
 level of the main factor.}

\item{rg}{A list (or \code{NULL}) containing the exponents \code{c(r, g)} of the
weights
  \eqn{w(x) = x^r (1-x)^g}. Both exponents need to be natural numbers including 0.
 Default is \code{list( c(0, 0) )} corresponding to the log-rank weight.}
}
\description{
The function \code{casanova} calculates the Wald-type statistic based on the
combination of differently weighted Nelson-Aalen-type integrals. Respective p-values
are obtained by a \eqn{\chi^2}-approximation and a permutation approach, respectively.
}
\details{

}
\examples{
library("survival")
data(veteran)
out <- casanova(formula ="time ~ trt*celltype",event = "status",
 data = veteran)

## Detailed informations:
summary(out)

}
\references{
Ditzhaus, M., Janssen, A. and Pauly, M. (2020). Permutation inference in factorial survival designs with the
          CASANOVA. ArXiv preprint (arXiv:2004.10818).
}
