% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FACTORIAL_FUN.R
\name{func_test}
\alias{func_test}
\title{Inference test for factorial designs in survival models}
\usage{
func_test(
  formula,
  event = "event",
  data = NULL,
  nperm = 10000,
  alpha = 0.05,
  cross = TRUE,
  nested.levels.unique = FALSE,
  rg = list(c(0, 0))
)
}
\arguments{
\item{formula}{A model \code{formula} object. The left hand side contains the time variable and the right
hand side contains the factor variables of interest. An interaction term must be specified.}

\item{event}{Name of the response variable}

\item{data}{A data.frame, list or environment containing the variables in formula. The
default option is \code{NULL}.}

\item{nperm}{The number of permutations used for calculating the permuted p-value.
The default option is 10000.}

\item{alpha}{A number specifying the significance level; the default is 0.05.}

\item{cross}{logical. Should the weight correspondng to crossing hazards be included?
The default is \code{TRUE}.}

\item{rg}{A list (or \code{NULL}) containing the exponents \code{c(r, g)} of the directions
 \eqn{w(x) = x^r (1-x)^g}. Both exponents need to be natural numbers including 0.
Default is \code{list( c(0, 0) )} corresponding to proportional hazards.}
}
\description{
The function names calculates the p-values of the multiple-direction logrank test based
on the \eqn{\chi^2}-approximation and the permutation approach.
}
\examples{
library("survival")
data(veteran)
out <- func_test(formula ="time ~ trt*celltype",event = "status",
 data = veteran, nperm = 1000, alpha = 0.05)

## Detailed informations:
summary(out)

}
\references{
Ditzhaus, M., Titel und so (Theory)

Ditzhaus Titel und so (practical paper)
}
